create app layout
